// src/app/modules/shared/model/register/user.model.ts
export class User {
  userId?: string;
  username?: string;
  userType?: number;
  email?: string;
  password?: string;
  address?: string;
  city?: string;
  state?: string;
  postcode?: number;
  country?: string;
  dob?: Date;
  gender?: string;
  fullName?: string;
  mobileNo?: string;
  race?: string;
  religion?: string;
  startDate?: Date;
  position?: string;
  profileImg?: string;
  profileImgPath?: string;
  createdBy?: string;
  createdDT?: Date;
}